requirements.txt
Flask-Bcrypt==0.7.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
marshmallow-sqlalchemy==0.26.1
PyJWT==2.1.0
pytest==6.2.5 